move enum and rename response class
